+++ inc/constants/game_header.h
@@ -10,11 +10,13 @@ enum GC_GAME
 #define GCL_INGAME_MAP_WIDTH 20
 #define GCL_INGAME_MAP_HEIGHT 17
 #define GCL_INGAME_BUILDING_SIZE 1
+#define GCL_INGAME_MONSTER_NEST_SIZE 3
 #define GCL_INGAME_FRAMERATE 30
 
 #define GCCS_INGAME_MAP_WIDTH 54
 #define GCCS_INGAME_MAP_HEIGHT 32
 #define GCCS_INGAME_BUILDING_SIZE 2
+#define GCCS_INGAME_MONSTER_NEST_SIZE 5
 #define GCCS_INGAME_FRAMERATE 30
 
 struct GameHeader
@@ -25,6 +27,7 @@ struct GameHeader
     int ingameMapHeight;
     int ingameBuildingSize;
     int ingameFramerate;
+    int ingameMonsterNestSize;
 };
 
 extern GameHeader g_game;