+++ inc/ingame/ingame_level_definition.h
@@ -5,10 +5,14 @@
 #include "constants/game_header.h"
 #include "constants/tile_type.h"
 
+#include <tuple>
+#include <vector>
+
 struct IngameLevelDefinition
 {
     IngameLevelDefinition() : tiles(g_game.ingameMapHeight, g_game.ingameMapWidth) {}
 
     vector2d<TILE_TYPE> tiles;
+    std::vector<std::tuple<int, int, bool>> monsterNests;
     int orbX, orbY;
 };