+++ inc/ingame/ingame_pathfinder.h
@@ -5,6 +5,7 @@
 #include "constants/status.h"
 #include "constants/tile_type.h"
 
+#include "entity/monster_nest.h"
 #include "entity/orb.h"
 #include "entity/path_edge_node.h"
 
@@ -15,12 +16,14 @@
 
 class IngameCore;
 class IngameMap;
+class IngameStructureController;
 struct IngameLevelDefinition;
 
 class IngamePathfinder
 {
 private:
     std::vector<PathEdgeNode> pathEdges;
+    std::vector<MonsterNest>& monsterNests;
     Orb& orbNode;
 
     static void calculateDistanceToNode(
@@ -29,7 +32,8 @@ private:
     void recalculateNodePaths(IngameMap& map, MonsterNode* pNode);
 
 public:
-    IngamePathfinder(IngameMap& map, IngameLevelDefinition& level);
+    IngamePathfinder(
+        IngameMap& map, IngameStructureController& structure, IngameLevelDefinition& level);
 
     void recalculatePaths(IngameMap& map);