+++ src/ingame/ingame_building_controller.cpp
@@ -89,8 +89,8 @@ void IngameBuildingController::render(struct _fbg* pFbg, const Window& window) c
 
     if (orb.isBroken())
     {
-        fbg_rect(pFbg, (orb.x - 0.5f * g_game.ingameBuildingSize) * scale + window.x,
-            (orb.y - 0.5f * g_game.ingameBuildingSize) * scale + window.y,
+        fbg_rect(pFbg, (orb.nodeX - 0.5f * g_game.ingameBuildingSize) * scale + window.x,
+            (orb.nodeY - 0.5f * g_game.ingameBuildingSize) * scale + window.y,
             scale * g_game.ingameBuildingSize, scale * g_game.ingameBuildingSize, 0x08, 0x08, 0x20);
     }
 }